requirements.txt
playwright
python-dotenv
python-dateutil